Version2++hamza

+++ src/main/java/com/civil/project/service/MarginaleArJugeNaissanceServiceImpl.java
@@ -1,11 +1,11 @@
 package com.civil.project.service;
 
 import com.civil.project.dao.JugeNaissanceMargArRep;
-import com.civil.project.dao.JugeNaissanceMargFrRep;
-import com.civil.project.dao.JugeNaissanceRegistreRep;
 import com.civil.project.entity.MargJugeNaissAr;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -19,8 +19,9 @@ public class MarginaleArJugeNaissanceServiceImpl implements MarginaleArJugeNaiss
 
     //ajouter marginal arabic de juge de naissance
     @Override
-    public void AjouterOuModifierMargAr(MargJugeNaissAr margJugeNaissAr) {
+    public MargJugeNaissAr AjouterOuModifierMargAr(MargJugeNaissAr margJugeNaissAr) {
         jugeNaissanceMargArRep.save(margJugeNaissAr);
+        return margJugeNaissAr;
     }
 
     //afficher les marginales arabic de juge de naissance
@@ -32,12 +33,21 @@ public class MarginaleArJugeNaissanceServiceImpl implements MarginaleArJugeNaiss
     //supprimer marginal arabic de juge de naissance
     @Override
     public void SupprimerMargAr(int idMargeJuge) {
-        jugeNaissanceMargArRep.deleteById(idMargeJuge);
+        if(MargArById(idMargeJuge)!=null) {
+            jugeNaissanceMargArRep.deleteById(idMargeJuge);
+        }
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND,
+                String.format("il n'y a aucun marginal de juge de naissance avec l'id %d  ",idMargeJuge));
+
     }
 
     //chercher le marginal arabic de juge de naissance a partir de id acte de juge de naissance
     @Override
     public List<MargJugeNaissAr> MargArByIdActeJugeNaissance(int id) {
+        if(jugeNaissanceMargArRep.findByIdActeJugeNais(id).isEmpty())
+        {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,String.format("il n'y a aucun marginal lie avec l'id %d d'acte de juge de naissance ",id));
+        }
         return jugeNaissanceMargArRep.findByIdActeJugeNais(id);
     }
 
@@ -53,7 +63,7 @@ public class MarginaleArJugeNaissanceServiceImpl implements MarginaleArJugeNaiss
             margJugeNaissAr=resultat.get();
         }
         else {
-            throw new RuntimeException("Registre non trouve");
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,"Marginale de juge de naissance arabe introuvable");
         }
         return margJugeNaissAr;
     }
